Fixed Charts Firefox issues (#971)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

It was found two major bugs of our charts in Firefox. 
1. Chart dots (e.g. for Line chart) were not displayed at all.
2. Appearing animations were broken.

Dots were not displayed due to some Firefox bug that doesn't accept `r`
property from styles and require `r` property to be set as attribute.

I totally doesn't know why animations were working in Chrome – it was
made as weird as possible. I have rebuild it with much robust approach
that work fine in Chrome, Safari and Firefox.

## How has this been tested?

Manual tests in Chrome, Safari and Firefox

## Screenshots:

Before fix:

![image](https://github.com/semrush/intergalactic/assets/31261408/15fe5c0e-ae31-4582-a013-7d82ec725027)

After fix:

![image](https://github.com/semrush/intergalactic/assets/31261408/a84cde98-4346-41d6-b659-2ba03310948d)

Broken animation:


https://github.com/semrush/intergalactic/assets/31261408/5a9b59c1-c627-4e72-9481-5880c1ca3ebf


## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.