[popper] added visual focus ring on trigger when focus is got by focus catcher (#989)
## Motivation and Context

Currently when focus on page is on the popper trigger focus catcher, the
focus ring is not visible.
I have added context that enabled components that use
keyboardFocusEnhance inside of popper trigger to highlight when some of
trigger focus catchers are focused by keyboard interaction.

## How has this been tested?

Not sure how to add tests for that, ideas are welcome.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [x] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.