[popper] fixed that after closing popper with interaction `click` focus is returned to the trigger instead of special focus catch element (#1009)
## Motivation and Context

We have focus catch element that catches focus after closing popper.
It's placed right next to the trigger and works like a charm for for
`interaction=hover` и `interaction=focus` but extremely weird for
`interaction=click`.

After returning focus it looks like trigger is focused but keyboard
pressing doesn't work.

Also I found issue with `triggerRef` – instead of `useRef` it was using
`createRef` that has no sense in functional components.

## How has this been tested?

Manual testing showed that everything works good.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.