[outside-click, utils] fixed outside clicks handling in shadow dom (#1042)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context


We recently moved all website examples into Shadow DON while all
components was never tested to work in Shadow DOM. It was found that
https://developer.semrush.com/intergalactic/utils/popper/popper render
function examples closes immediately after opening.

The example was broken because OutsideClick tries to get what element is
clicked and in case of using shadow dom elements it's the shadow dom
container – the whole playground. So, for outside click component every
click was the outside one.

## How has this been tested?

The change is extremely simple – I've just made it to support shadow DOM
with full backward capability.
So I think we do not need shadow DOM tests for now. 

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.