[86bxd8qjz] added space keydown handler alongside all enter keydown handlers (#1096)
## Motivation and Context

We have a lot of interactive elements that listen to Enter keydown and
call the click handlers. It's also expected for them to react on Space
keydowns.

I've checked out all placed in code where keydown

## How has this been tested?

I've manually tested every updated component. Nothing seems to be risky
or require additional unit tests (over existing ones for Enter key).

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [x] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.