[tooltip] fixed reading popper content twice and added popper id to trigger children render func params (#1094)
## Motivation and Context

This PR resolve following issues:
1. When trigger is focused, tooltip content was read twice by assistive
technologies. I've fixed by setting `aria-live` on a `display:
contents;` element that is not focusable (surprisingly found that it
helps after reading last paragraph of [this
discussion](https://stackoverflow.com/a/75189223/6025068)).
2. `aria-describedby` is always applied to the trigger container. It's
good in most cases but sometimes it must be applied to a focusable
element inside of trigger (e.g. Link or Interactive icon). So I've added
popperId param to render function of trigger children. Checkout the
added examples to see it.

## How has this been tested?

Manual testing with Voice Over.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [x] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.