[tag] if tag contains `Tag.Close`, tag will not be focusable, instead, `Tag.Text` will be focusable (#1114)
## Motivation and Context

Our tags close icon is annotated as "tap to remove tag <tag-text>".
So, the whole tag is annotated as "<tag-text> tap to remove tag
<tag-text>". To deal with it, I've added a Tag root state, that depends
on `Tag.Close` mount/unmount and decides which elements gets
`tabIndex={0}` – `Tag` or `Tag.Text`.

Now VO reads just "<tag-text"> on tag focus and than "tap to remove tag
<tag-text>" on close icon.

## How has this been tested?

Manually and with new added unit tests.

## Screenshots (if appropriate):

See changed files.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.