[86bxkev49][base-trigger] fixed that filter trigger was setting aria attributes to wrapper instead of inner trigger (#1123)
## Motivation and Context

Our Filter trigger was found to not be accessible cause it was setting
aria attributes on outer wrapper instead of inner trigger. I've fixed
it. Solution is not perfect cause we need to add deep component
customization but it's good enough just to make it accessible.

## How has this been tested?

Manual testing, soon I will enable a11y tests.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.