[86bxcxzxy][date-picker] a11y fixes (#1142)
## Motivation and Context

This PR contains a set of critical a11y fixes of date picker components.

1. I have rewrote range comparator keyboard handler to handle popper
keydowns.
2. I added `unitRefs` – it allows to focus on `CalendarUnit` when user
navigates with arrows.`
3. In all date picker poppers I removed `event.target` check as it was
preventing keyboard handling when calendar unit is focused.
4. I added `abbr` attribute on calendar grid for all "short" forms of
date terms. The `propsForElement` awful regexp was not allowing to set
it, so I modified the regexp.
5. I added aria-live region that produces description of what happens
when user selects date ranges.
6. I made date-picker popper elements focusable and not aria-hidden.
Also I changed focus order to focus on periods and apply/clear button
first.
7. Also I found that in range comparators periods presets had no padding
– i've added it.
8. I've removed `aria-expanded` attribute from input trigger as it was
violating a11y automatic checks. To ensure with it, I've updated a11y
test from using old button trigger and replaced it with input trigger.


## How has this been tested?

With existing and updated unit tests. 

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [x] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.