[dropdown-menu] footer focus a11y (#1145)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

After #1111 it was found that focusable elements inside of dropdown-menu
footer are not very accessible. It's very hard to get to them with tab
key navigation.

I have added additional check that disabled focus lock if focus goes out
of item but is kept inside of dropdown popper.

## How has this been tested?

Manually on wide range of cases and with unit test on the main case –
tab navigation goes from one focusable to another.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.