[global] some fixes in publish release scripts (#1188)
`simple-git` `tags` command return list of tags, sorted by semantic
version number by default. So, as currentTag in release we have latest
prerelease tag of this version. Also, sort by `creatordate` doesn't
work, because for prerelease and release we'll have the same dates, so,
we'll get prerelease too.
I've changed `getCurrentTag` method to use `git describe` and fixed
version in message to slack and fixed a list of tasks to close.

Also, I've changed logic of generating changelogs and publishing for
`intergalactic` package.