[86bxq08pz][data-table] added `renderRows` prop (#1187)
## Motivation and Context

Our virtual rendering mechanism in data table is pretty simple. It's
restricted to tables with rows of fixed height.

Supporting virtual scrolling of rows with variating height is too
complex for our data table and has too many pitfalls.

So, `rowsRendering` prop was added. It allows to redefine rows rendering
and customize virtual scroll.

## How has this been tested?

Manually based on added example.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue).
- [x] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.

---------

Co-authored-by: ilia.brauer <ilia.brauer@semrush.com>