[86bxwjwjc][utils] use React.useId in useUID when it's avaialble (#1196)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

On one of existing project hydration errors were faced. It was caused by
different `uid` generated by `useUID` hook on the client and on the
server. I made it to use React.useId when it's available (we want to
support older versions of React).

## How has this been tested?

I've modified code of library in node_modules of that project and
ensured that hydration errors have gone.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.