[86bxxhdzg][utils, popper] made focus lock to work in popper with disabled portal (#1218)
## Motivation and Context

If developer disables popper portal, popper prevents blur event to
bubble of it. It's critical for work of focus lock, so I've added
synthetic events in focus lock and made popper to forward it when it
prevents specific event from bubbling.

## How has this been tested?

I have added focus lock browser tests and covered disabledPortal caase.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.