[86bxwjwjc][portal] removed initial mount node (#1238)
## Motivation and Context

Commonly react frameworks
[recommend](https://github.com/vercel/next.js/blob/canary/examples/with-portals/components/ClientOnlyPortal.js)
to delay mount of portals with state + effect to require a state change
to prevent ssr hydration errors (when we have `null` on server but
portal on client).

IDK why we didn't followed this rule of thumb before and made it to
return `getNodeByRef(container)` initially on server. It was causing
hydration errors at least in nextjs and astro.

## How has this been tested?

Tooltip was causing hydration errors in an astro project. After this
update all errors gone.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.