[86by5t73d][popper] fixed that Select and similar components with disabled portal were not working properly being wrapped into `label` tag (#1273)
## Motivation and Context

Long time ago Select and other components were not supporting label tag
clicks handling. Recently we fixed it. On the other had some of
developers were wrapping Select's into labels without intention to get
correct clicks handling and just for a11y tree labeling purpose.

Last case was broken by our update.
In this PR I'm adding additional checks. As we can't check if click
comes from label tag I have added following check:
if trigger comes in window of 100ms after popper click and popper is a
children of some of a label that is related to trigger, we should ignore
this click.

## How has this been tested?

With added browser tests.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.