[86bxmp5xu] Ellipsis with links (#1278)
## Motivation and Context

It was found that 

```
<Ellipsis tag={Link} trim='middle' href='https://developer.semrush.com/intergalactic/'>
  https://developer.semrush.com/intergalactic/
</Ellipsis>
```

Renders as not link.

```
<Link tag={Ellipsis} trim='middle' href='https://developer.semrush.com/intergalactic/'>. https://developer.semrush.com/intergalactic/
</Link>
```

Doesn't renders at all.

The first case I fixed to make it render like a normal link. The second
case I made to render as non-link – now it makes DX not so frustrating
and allows developers to find a better approach of the component use.

## How has this been tested?

Mainly manually on many cases and on a few cases with unit tests.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.