[86by0byqr][select] fixed multiselect a11y (#1282)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

Multiselect wasn't annotated properly if it used with `options` prop
because default options rendered was adding aria-selected without taking
into account that `value` may be an array of selected options.

Also I've fixed a11y linting – select.option.checkbox should not have
any roles because it's just a visual representation of multiselect
option.

## How has this been tested?

Manually and with unit test.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.