[86byj6hrj][popper] fixed that poppers with `interaction=hover` were not opening on touch devices (#1337)
## Motivation and Context

Recently additional check for mouseenter event was added that checks
that mousemove event had happened just before.

It doesn't work on touchdevices – despite triggering mouseenter, it
doesn't produces mousemove before the mouseenter.

To fix it, I've added touchstart to the list of events that trigger.

Also I've fixed some tests names and enabled touch support in browser
tests.

## How has this been tested?

Manually and with added browser test.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.

Co-authored-by: ilia.brauer <ilia.brauer@semrush.com>