[website] update a11y page for WidgetEmpty (#1355)
Small update to Accessibility documentation for WidgetEmpty component. 

I was thinking whether we should remove the entire 'What component has'
section (as we were discussing), but it seems better to leave it because
it shows what the component already has and what our users can check and
test while using this component in their projects.