[86bz07bvp][popper] fixed combobox pattern with mouse interactions (#1383)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

It was found that after selecting an item in combobxed select with mouse
the popper reopens again. While it's not totally making comboboxes
unusable, it looks very buggy.

Adding `if (!this.asProps.visible) {` seems to be very reasonable but
I'm not sure what was the idea behind adding `else if (now <
this.state.ignoreTriggerFocusUntil + ignoringDuration) {` conditional
branching. Probably it was added "just to be safe" but for now I can't
find any useful case that was broken by removing it.

## How has this been tested?

I have tested may dropdown examples on the website manually and didn't
found any bugs.
Also I've created a new browser test.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [x] I have added new unit tests on added of fixed functionality.