[86bzughj1][utils] made focus lock not to fall into iframes (#1623)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

When browser focus is comming to iframe, we are loosing control of it. I
tried to use `document.hasFocus()` api or react to events before browser
focus is controlled by iframe, but found that it looks impossible. To
resolve the issue, I've added process of adding invisible focusable
elements that prevent browser focus from reaching iframe.

## How has this been tested?

I found no way to check which window has focus, root document always
gets true from `document.hasFocus()`. So, only manually.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [ ] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.