[86c0qcv36][docs] Updated DragAndDrop Design and A11y pages (#1712)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

1. Updated the A11y page. Didn't include some questionable attributes
(like tabindex=0 in DropZone) that I hope we will remove/fix soon.
2. Completely overhauled the Design page. A lot of information repeated,
so I decided to make the structure more linear and remove repetitions.

There's one questionable moment. The whole guide was mostly written
(even before my edits) as if objects can only be grabbed by the drag
area/indicator. BUT:

1. This is not how it works now (we have a task to change it in the
future, but the docs are updated right now).
2. Even the guide itself has a couple of paragraphs that look like
entire objects can be grabbed. I commented them out for now. If we meant
that both cases are possible, then the distinction should be described
more clearly (and I need more info for that).

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [x] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [ ] Unit tests are not broken.
- [ ] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.