[UIK-2735][docs] Labels and instructions (#1942)
<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

This PR adds missing labels and improves existing ones for various
controls in our docs and examples.

Some highlights:

1. The only component change: fixed the name of a token in Button
styles.
2. Dot: added aria-describedby to examples, but there still are a11y
issues, I created a separate task.
3. ScrollArea: added aria-labels to scroll containers in examples, so
when you focus a container it's clear what it is. Except for the first
one, I can't add the label without expanding the component, and I wasn't
sure it would be appropriate for the basic example.
4. Checkbox and Radio playgrounds: renamed "children" to "label".
5. InputTags playground: removed the text property because it didn't
seem useful.
6. SpinContainer playground: changed the structure of inputs to be
accessible when "loading" is off.
7. Spin playground: removed "centered" prop because its behavior didn't
make sense, created a separate task. Also renamed the TextRight
property.
8. Switch: added icon changing between on/off states logic in Playground
and examples, kept only one addon in Playground and renamed it to
"Label" for consistency with Radio & Checkbox. Created a task to add the
"label" property instead addons.
9. Tag: rewrote playground to use TagContainer and Close correctly.
10. Charts: removed some useless Flex props from Playgrounds, renamed
props to be clearer and shorter. Renamed "lines" and "bars" to
"categories" (and vice versa) wherever needed (not everywhere, mostly in
examples with legend).
11. Box: removed the duplicate sandbox. I don't know why it was there.
12. Popper: Render functions example: added returning focus to 1st
button after closing the popper, removed the duplicate sandbox.
Controlled and uncontrolled example: replaced triggers with our Buttons.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected).
- [x] Nice improve.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] I have updated the documentation accordingly or it's not required.
- [x] Unit tests are not broken.
- [x] I have added changelog note to corresponding `CHANGELOG.md` file
with planned publish date.
- [ ] I have added new unit tests on added of fixed functionality.

---------

Signed-off-by: Sheila <166654065+sheila-semrush@users.noreply.github.com>